UefiCpuPkg/RegisterCpuFeaturesLib: Fix build failure.
Build UefiCpuPkg with below configuration:
Architecture(s)  = IA32
Build target     = NOOPT
Toolchain        = VS2015x86

Below error info shows up:
DxeRegisterCpuFeaturesLib.lib(CpuFeaturesInitialize.obj) :
error LNK2001: unresolved external symbol __allmul

Valid mDependTypeStr type only have 5 items, use UINT32 type cast
to fix this error.

Cc: Dandan Bi <dandan.bi@intel.com>
Cc: Ruiyu Ni <ruiyu.ni@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1

+++ UefiCpuPkg/Library/RegisterCpuFeaturesLib/CpuFeaturesInitialize.c
@@ -517,7 +517,7 @@ DumpRegisterTableOnProcessor (
         DebugPrintErrorLevel,
         "Processor: %d: Semaphore: Scope Value: %s\r\n",
         ProcessorNumber,
-        mDependTypeStr[MIN (RegisterTableEntry->Value, InvalidDepType)]
+        mDependTypeStr[MIN ((UINT32)RegisterTableEntry->Value, InvalidDepType)]
         ));
       break;
